refactor(api): move certification-center-creation validator
in organizational-entities
