+++ api/tests/organizational-entities/unit/domain/validators/certification-center-creation.validator.test.js
@@ -1,6 +1,6 @@
-import * as certificationCenterCreationValidator from '../../../../lib/domain/validators/certification-center-creation-validator.js';
-import { EntityValidationError } from '../../../../src/shared/domain/errors.js';
-import { expect } from '../../../test-helper.js';
+import * as certificationCenterCreationValidator from '../../../../../src/organizational-entities/domain/validators/certification-center-creation.validator.js';
+import { EntityValidationError } from '../../../../../src/shared/domain/errors.js';
+import { expect } from '../../../../test-helper.js';
 
 const MISSING_VALUE = '';
 
@@ -10,10 +10,10 @@ function _assertErrorMatchesWithExpectedOne(entityValidationErrors, expectedErro
   expect(entityValidationErrors.invalidAttributes[0]).to.deep.equal(expectedError);
 }
 
-describe('Unit | Domain | Validators | certification-center-validator', function () {
+describe('Unit | Organizational Entities | Domain | Validators | certification-center-validator', function () {
   describe('#validate', function () {
     context('when validation is successful', function () {
-      it('should not throw any error', function () {
+      it('does not throw any error', function () {
         // given
         const certificationCenterCreationParams = { name: 'ACME', type: 'PRO' };
 
@@ -24,7 +24,7 @@ describe('Unit | Domain | Validators | certification-center-validator', function
 
     context('when certification-center data validation fails', function () {
       context('on name attribute', function () {
-        it('should reject with error when name is missing', function () {
+        it('rejects with error when name is missing', function () {
           // given
           const expectedError = {
             attribute: 'name',
@@ -45,7 +45,7 @@ describe('Unit | Domain | Validators | certification-center-validator', function
           }
         });
 
-        it('should reject with error when name is longer than 255 characters', function () {
+        it('rejects with error when name is longer than 255 characters', function () {
           // given
           const expectedError = {
             attribute: 'name',
@@ -68,7 +68,7 @@ describe('Unit | Domain | Validators | certification-center-validator', function
       });
 
       context('on type attribute', function () {
-        it('should reject with error when type is missing', function () {
+        it('rejects with error when type is missing', function () {
           // given
           const expectedError = [
             {
@@ -97,7 +97,7 @@ describe('Unit | Domain | Validators | certification-center-validator', function
           }
         });
 
-        it('should reject with error when type value is not SUP, SCO or PRO', function () {
+        it('rejects with error when type value is not SUP, SCO or PRO', function () {
           // given
           const expectedError = {
             attribute: 'type',
@@ -117,7 +117,7 @@ describe('Unit | Domain | Validators | certification-center-validator', function
 
         // eslint-disable-next-line mocha/no-setup-in-describe
         ['SUP', 'SCO', 'PRO'].forEach((type) => {
-          it(`should not throw with ${type} as type`, function () {
+          it(`does not throw with ${type} as type`, function () {
             // given
             const certificationCenterCreationParams = { name: 'ACME', type };
 
@@ -129,7 +129,7 @@ describe('Unit | Domain | Validators | certification-center-validator', function
       });
 
       context('on externalId attribute', function () {
-        it('should reject with error when externalId is longer than 255 characters', function () {
+        it('rejects with error when externalId is longer than 255 characters', function () {
           // given
           const expectedError = [
             {
@@ -156,7 +156,7 @@ describe('Unit | Domain | Validators | certification-center-validator', function
         });
       });
 
-      it('should reject with errors on all fields (but only once by field) when all fields are missing', function () {
+      it('rejects with errors on all fields (but only once by field) when all fields are missing', function () {
         // given
         const certificationCenterCreationParams = {
           name: MISSING_VALUE,
